python-common, orch: Move ServiceSpec tests to python-common
